+++ javaparser-core/src/main/java/com/github/javaparser/ast/validator/Validator.java
@@ -8,7 +8,7 @@ import java.util.function.BiConsumer;
  * A validator that can be run on a node to check for semantic errors.
  * It is fully up to the implementor how to do this.
  */
-public interface Validator extends BiConsumer<Node, ProblemReporter> {
+public interface Validator extends TypedValidator<Node> {
     /**
      * @param node the node that wants to be validated
      * @param problemReporter when found, validation errors can be reported here
